Reworked Shield Pushback
