+++ fighters/common/src/status/guard/guard_damage.rs
@@ -7,7 +7,7 @@ unsafe extern "C" fn status_pre_guarddamage(fighter: &mut L2CFighterCommon) -> L
         fighter.module_accessor,
         SituationKind(*SITUATION_KIND_GROUND),
         *FIGHTER_KINETIC_TYPE_MOTION,
-        *GROUND_CORRECT_KIND_GROUND_CLIFF_STOP as u32,
+        *GROUND_CORRECT_KIND_GROUND as u32,
         GroundCliffCheckKind(*GROUND_CLIFF_CHECK_KIND_NONE),
         true,
         *FIGHTER_STATUS_WORK_KEEP_FLAG_GUARD_DAMAGE_FLAG,
@@ -83,7 +83,7 @@ unsafe extern "C" fn sub_ftstatusuniqprocessguarddamage_initstatus_inner(fighter
         shield_stiff_frame = shield_stiff_frame_max;
     }
     // println!("final shield_stiff_frame: {}", shield_stiff_frame as i32);
-    let object_id = WorkModule::get_int(fighter.module_accessor, *FIGHTER_STATUS_GUARD_DAMAGE_WORK_INT_OBJECT_ID);
+    // let object_id = WorkModule::get_int(fighter.module_accessor, *FIGHTER_STATUS_GUARD_DAMAGE_WORK_INT_OBJECT_ID);
     // if object_id != *BATTLE_OBJECT_ID_INVALID {
         // capture!(fighter, MA_MSC_CMD_CAPTURE_SET_IGNORE_OBJECT_ID, object_id);
         let mut invalid_capture_frame = shield_stiff_frame;
@@ -198,20 +198,34 @@ unsafe extern "C" fn sub_ftstatusuniqprocessguarddamage_initstatus_inner(fighter
         let shield_setoff_speed_mul = WorkModule::get_param_float(fighter.module_accessor, hash40("common"), hash40("shield_setoff_speed_mul"));
         let shield_setoff_speed_max = WorkModule::get_param_float(fighter.module_accessor, hash40("common"), hash40("shield_setoff_speed_max"));
         let shield_lr = -WorkModule::get_float(fighter.module_accessor, *FIGHTER_STATUS_GUARD_DAMAGE_WORK_FLOAT_SHIELD_LR);
-        let setoff_speed = (shield_setoff_speed_mul * shield_stiff_frame * shield_lr).clamp(-shield_setoff_speed_max, shield_setoff_speed_max);
+        // let setoff_speed = (shield_setoff_speed_mul * shield_stiff_frame * shield_lr).clamp(-shield_setoff_speed_max, shield_setoff_speed_max);
         // println!("setoff_speed: {}", setoff_speed);
-        let setoff_mul_indirect = if sv_battle_object::category(object_id as u32) != *BATTLE_OBJECT_CATEGORY_FIGHTER {
-            0.5
+        let mut setoff_speed = (shield_power * shield_setoff_speed_mul + 0.5) * shield_lr;
+        // println!("shield_power {} * mul {} + 0.5", shield_power, shield_setoff_speed_mul);
+        // println!("setoff_speed: {}", setoff_speed);
+
+        fighter.clear_lua_stack();
+        lua_args!(fighter, FIGHTER_KINETIC_ENERGY_ID_DAMAGE);
+        let current_x = sv_kinetic_energy::get_speed_x(fighter.lua_state_agent);
+        if current_x.signum() == setoff_speed.signum() {
+            let count = WorkModule::get_int(fighter.module_accessor, *FIGHTER_STATUS_GUARD_DAMAGE_WORK_INT_DAMAGE);
+            let mul = if count > 0 {
+                0.2
+            }
+            else {
+                1.0
+            };
+
+            // println!("combined! current {} + setoff {} = {}", current_x, setoff_speed * mul, current_x + (setoff_speed * mul));
+            setoff_speed = current_x + (setoff_speed * mul);
         }
-        else {
-            1.0
-        };
+
         sv_kinetic_energy!(
             reset_energy,
             fighter,
             FIGHTER_KINETIC_ENERGY_ID_DAMAGE,
             ENERGY_STOP_RESET_TYPE_GUARD_DAMAGE,
-            setoff_speed * setoff_mul_indirect,
+            setoff_speed.clamp(-shield_setoff_speed_max, shield_setoff_speed_max),
             0.0,
             0.0,
             0.0,
@@ -496,6 +510,16 @@ unsafe extern "C" fn status_guarddamage_main(fighter: &mut L2CFighterCommon) -> 
     0.into()
 }
 
+#[skyline::hook(replace = L2CFighterCommon_status_guard_damage_main_common_air)]
+unsafe extern "C" fn status_guard_damage_main_common_air(fighter: &mut L2CFighterCommon) -> L2CValue {
+    if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_AIR {
+        // fighter.change_status(FIGHTER_STATUS_KIND_DAMAGE_FALL.into(), false.into());
+        fighter.change_status(FIGHTER_STATUS_KIND_MISS_FOOT.into(), false.into());
+        return 1.into();
+    }
+    0.into()
+}
+
 #[skyline::hook(replace = L2CFighterCommon_sub_ftStatusUniqProcessGuardDamage_execStatus_common)]
 unsafe extern "C" fn sub_ftstatusuniqprocessguarddamage_execstatus_common(fighter: &mut L2CFighterCommon) {
     if !WorkModule::is_flag(fighter.module_accessor, *FIGHTER_STATUS_GUARD_ON_WORK_FLAG_JUST_SHIELD) {
@@ -663,6 +687,7 @@ fn nro_hook(info: &skyline::nro::NroInfo) {
             status_guarddamage_common,
             sub_guarddamageuniq,
             status_guarddamage_main,
+            status_guard_damage_main_common_air,
             sub_ftstatusuniqprocessguarddamage_execstatus_common,
             sub_ftstatusuniqprocessguarddamage_execstop_common,
             fighterstatusuniqprocessguarddamage_leave_stop,
